Extract tool path overlay
