+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+  "chat.tools.terminal.autoApprove": {
+    "npm.cmd": true
+  }
+}
