+++ src/renderer/src/features/canvas/index.ts
@@ -9,3 +9,4 @@ export * from "./utils/geometry";
 export * from "./components/BedLayer";
 export * from "./components/GridLayer";
 export * from "./components/SelectionOverlay";
+export * from "./components/ToolpathOverlay";
